Bug #1240686 mmc: sprd-sdhcr10: Fix compile error
[root cause ] data type error
[changes ] code
[side effects] no
[self test ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]: 8541E
[test case]:system boot ok
[reviewers ] billows wu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Fix compile error. Cast to 'void *' from smaller integer
type 'int' is error. the size of 'void *' is bigger
than 'int' in 64 bit system.

+++ drivers/mmc/host/sprd-sdhcr10.c
@@ -1360,7 +1360,7 @@ static int sprd_sdhc_get_cd(struct mmc_host *mmc)
 		return 1;
 	}
 
-	if (IS_ERR_VALUE(gpio_cd))
+	if (IS_ERR_VALUE((unsigned long)gpio_cd))
 		gpio_cd = 1;
 	mmiowb();
 	spin_unlock_irqrestore(&host->lock, flags);
